Alertas Reg/Edit/Del

+++ Frontend/src/views/borrarDatos/Empleado.jsx
@@ -1,10 +1,12 @@
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import axios from "axios";
+import Alert from '@mui/material/Alert'
 
 function Empleado() {
 
     const [rut, setRut] = useState("");
     const [exitoBorrarEmpleado, setExitoBorrarEmpleado] = useState("");
+    const [alertType, setAlertType] = useState("success")
 
     const handleOnChangeRut = (e) => {
         setRut(e.target.value)
@@ -13,24 +15,39 @@ function Empleado() {
     const handleSubmit = (e) => {
         e.preventDefault();
 
+        if (rut === "") {
+            setAlertType("warning")
+            setExitoBorrarEmpleado("Todos los campos son obligatorios")
+            return;
+        }
+
         axios.delete(`http://localhost:5000/empleados/delete`,{data:{rut_empleado:rut}})
             .then((response) => {
+            setAlertType("success")
             setExitoBorrarEmpleado("Empleado borrado exitosamente")
             console.log("Empleado borrado exitosamente", response.data)
             })
         .catch ((error) => {
+            setAlertType("error")
             setExitoBorrarEmpleado("Error al borrar el Empleado")
             console.error("Error al borrar el Empleado", error)
         });
     }
-        const ping = exitoBorrarEmpleado ? (
-            <div className='mt-3'>
-                <p>
-                    {exitoBorrarEmpleado}: <br/>
-                    {rut}
-                </p>
-            </div>
-        ): null;
+    useEffect(() => {
+        if (exitoBorrarEmpleado){
+            const timer = setTimeout(() => {
+                setExitoBorrarEmpleado(null);
+                setAlertType("success");
+            }, 7000);
+            return () => clearTimeout(timer);
+        }
+    }, [exitoBorrarEmpleado]);
+
+    const ping = exitoBorrarEmpleado ? (
+        <div className='mt-3'>
+            <Alert severity={alertType}>{exitoBorrarEmpleado}</Alert>
+        </div>
+    ): null;
 
     return(
         <>
